+++ packages/ckeditor5-dev-docs/lib/validators/index.js
@@ -5,29 +5,33 @@
 
 'use strict';
 
+const { utils } = require( '@ckeditor/typedoc-plugins' );
 const seeValidator = require( './see-validator' );
 const linkValidator = require( './link-validator' );
 const firesValidator = require( './fires-validator' );
-const overloadsValidator = require( './overloads-validator' );
 const moduleValidator = require( './module-validator' );
-const { getNode } = require( './utils' );
+const overloadsValidator = require( './overloads-validator' );
 
 /**
  * Validates the CKEditor 5 documentation.
  *
  * @param {Object} project Generated output from TypeDoc to validate.
  * @param {Object} typeDoc A TypeDoc application instance.
+ * @param {TypedocValidator} [options={}] A configuration object.
  * @returns {Boolean}
  */
 module.exports = {
-	validate( project, typeDoc ) {
+	validate( project, typeDoc, options = {} ) {
 		const validators = [
 			seeValidator,
 			linkValidator,
 			firesValidator,
-			overloadsValidator,
 			moduleValidator
 		];
+
+		if ( options.enableOverloadValidator ) {
+			validators.push( overloadsValidator );
+		}
 
 		typeDoc.logger.info( 'Starting validation...' );
 
@@ -39,7 +43,7 @@ module.exports = {
 
 		for ( const validator of validators ) {
 			validator( project, ( error, reflection ) => {
-				const node = getNode( reflection );
+				const node = utils.getNode( reflection );
 
 				errors.set( node, { error, node } );
 			} );