+++ packages/ckeditor5-dev-release-tools/lib/index.js
@@ -6,15 +6,18 @@
 'use strict';
 
 const releaseSubRepositories = require( './tasks/releasesubrepositories' );
+const preparePackages = require( './tasks/preparepackages' );
 const bumpVersions = require( './tasks/bumpversions' );
 const generateChangelogForSinglePackage = require( './tasks/generatechangelogforsinglepackage' );
 const generateChangelogForMonoRepository = require( './tasks/generatechangelogformonorepository' );
 const updateCKEditor5Dependencies = require( './tasks/updateckeditor5dependencies' );
+const updateDependenciesVersions = require( './utils/updatedependenciesversions' );
 const { getLastFromChangelog, getCurrent, getLastTagFromGit } = require( './utils/versions' );
 const { getChangesForVersion, getChangelog, saveChangelog } = require( './utils/changelog' );
 
 module.exports = {
 	releaseSubRepositories,
+	preparePackages,
 	bumpVersions,
 	generateChangelogForSinglePackage,
 	generateChangelogForMonoRepository,
@@ -24,5 +27,6 @@ module.exports = {
 	getLastTagFromGit,
 	getChangesForVersion,
 	getChangelog,
-	saveChangelog
+	saveChangelog,
+	updateDependenciesVersions
 };