+++ packages/typedoc-plugins/tests/event-inheritance-fixer/fixtures/mixed-class.ts
@@ -1,0 +1,12 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+import type { InterfaceC } from './interface-c';
+
+/**
+ * @module fixtures/mixed-class
+ */
+
+export class MixedClass implements InterfaceC {}