+++ packages/typedoc-plugins/tests/interface-augmentation-fixer/index.js
@@ -8,6 +8,7 @@ const glob = require( 'fast-glob' );
 const TypeDoc = require( 'typedoc' );
 
 const utils = require( '../utils' );
+const { plugins } = require( '../../lib' );
 
 describe( 'typedoc-plugins/interface-augmentation-fixer', function() {
 	this.timeout( 10 * 1000 );
@@ -33,8 +34,8 @@ describe( 'typedoc-plugins/interface-augmentation-fixer', function() {
 			logLevel: 'Error',
 			entryPoints: files,
 			plugin: [
-				require.resolve( '@ckeditor/typedoc-plugins/lib/module-fixer' ),
-				require.resolve( '@ckeditor/typedoc-plugins/lib/interface-augmentation-fixer' )
+				plugins[ 'typedoc-plugin-module-fixer' ],
+				plugins[ 'typedoc-plugin-interface-augmentation-fixer' ]
 			],
 			tsconfig: utils.normalizePath( FIXTURES_PATH, 'tsconfig.json' )
 		} );